refactor: split HTTP request/response parsing out of PairingHTTPServer into its own Foundation-only file so the wire format can be tested without Network
